made the command loader independent I am trying to prepare to do this in TypeScript
